Add on click listener for recycler view

+++ app/src/main/java/com/knightua/notepadapp/room/entity/Note.kt
@@ -11,5 +11,5 @@ data class Note(
     @ColumnInfo(name = "date_of_creation") val dateOfCreation: Long?
 ) {
     @PrimaryKey(autoGenerate = true)
-    var id: Int = 0
+    var id: Long = 0
 }
